@@ -15,12 +15,12 @@
   },
   "dependencies": {
     "@mdi/js": "^7.2.96",
-    "@vueuse/core": "^10.2.1",
+    "@vueuse/core": "^10.3.0",
     "axios": "^1.4.0",
     "echarts": "^5.4.3",
     "lodash": "^4.17.21",
     "nprogress": "^0.2.0",
-    "pinia": "^2.1.4",
+    "pinia": "^2.1.6",
     "portal-vue": "^2.1.7",
     "vue": "^2.7.14",
     "vue-echarts": "^6.6.0",
@@ -33,11 +33,11 @@
     "@intlify/unplugin-vue-i18n": "^0.12.2",
     "@kingyue/rollup-plugin-modify": "^4.1.0",
     "@kingyue/vite-plugin-vue2-svg": "^0.6.0",
-    "@pinia/testing": "^0.1.2",
+    "@pinia/testing": "^0.1.3",
     "@rushstack/eslint-patch": "^1.3.2",
     "@testing-library/vue": "^5.9.0",
-    "@types/lodash": "^4.14.195",
-    "@types/node": "^18.16.19",
+    "@types/lodash": "^4.14.196",
+    "@types/node": "^18.17.2",
     "@types/nprogress": "^0.2.0",
     "@vitejs/plugin-legacy": "^4.1.1",
     "@vitejs/plugin-vue2": "^2.2.0",
@@ -45,31 +45,31 @@
     "@vue/eslint-config-typescript": "^11.0.3",
     "@vue/test-utils": "^1.3.6",
     "browserslist-to-esbuild": "^1.2.0",
-    "cypress": "^12.17.2",
-    "eslint": "^8.45.0",
+    "cypress": "^12.17.3",
+    "eslint": "^8.46.0",
     "eslint-plugin-cypress": "^2.13.3",
-    "eslint-plugin-vue": "^9.15.1",
+    "eslint-plugin-vue": "^9.16.1",
     "flush-promises": "^1.0.2",
     "jsdom": "^22.1.0",
     "msw": "^1.2.3",
     "postcss-load-config": "^4.0.1",
-    "postcss-preset-env": "^9.0.0",
-    "prettier": "^3.0.0",
+    "postcss-preset-env": "^9.1.0",
+    "prettier": "^3.0.1",
     "sass": "~1.32.13",
     "start-server-and-test": "^2.0.0",
-    "terser": "^5.19.1",
+    "terser": "^5.19.2",
     "ts-node": "^10.9.1",
     "typescript": "^5.1.6",
     "unplugin-auto-import": "^0.16.6",
     "unplugin-vue-components": "^0.25.1",
-    "vite": "^4.4.5",
-    "vite-plugin-inspect": "^0.7.33",
+    "vite": "^4.4.8",
+    "vite-plugin-inspect": "^0.7.35",
     "vite-plugin-pages": "^0.31.0",
     "vite-plugin-vue-layouts": "^0.8.0",
-    "vitest": "^0.33.0",
+    "vitest": "^0.34.1",
     "volar-plugin-vue-router": "^0.2.0",
     "vue-template-compiler": "^2.7.14",
-    "vue-tsc": "^1.8.5"
+    "vue-tsc": "^1.8.8"
   },
   "browserslist": [
     "> 1.3%",
